Revised Homepage Layout (News & Featured) (#119)
* feat: rearrange homepage

* feat: separate out the news/main-menu code and styles

includes featured, same method will be adopted when code for that is written

* campaigns: update hover style to bring consistency with the home page

* feat: add auto advance on news reel

includes pagination using indicator lights, tooltip changes, and adding a missing event dispatcher to mainmenu

* main-menu: move featured btn again

* refactor: prettier pass

* refactor: main-menu indentation

this screwed up prettier and made indentation adjustments by it *worse*. that's fixed here and now it's all Pretty! like it should be.

* main-menu: remove onFeaturedFlyoutPressed

not used and forgot to remove

* feat: better erroring on failed news retrieval

* fix: tooltip right arrow pointing in wrong direction

* feat: news nav arrows

* chore: nits

* addons: hide steam button for local addons

just slippin this with news-layout, dont really feel like making a whole separate thing for this